@@ -1,10 +1,23 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -12,7 +25,13 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.squareup.picasso.Picasso;
+
+import java.util.Objects;
 
 import arc.resource.calculator.adapters.ShowcaseResourceListAdapter;
 import arc.resource.calculator.db.DatabaseContract;
@@ -23,18 +42,6 @@ import arc.resource.calculator.util.AdUtil;
 import arc.resource.calculator.util.ExceptionUtil;
 import arc.resource.calculator.util.FeedbackUtil;
 
-/**
- * Copyright (C) 2016, Jared Stone
- * -
- * Author: Jared Stone
- * Title: A:RC, a resource calculator for ARK:Survival Evolved
- * -
- * Web: https://github.com/jaredstone1982/CraftingCalcARK
- * Email: jaredstone1982@gmail.com
- * Twitter: @MasterxOfxNone
- * -
- * This work is licensed under a Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International License.
- */
 public class DetailActivity extends AppCompatActivity
         implements ExceptionObserver.Listener {
     private static final String TAG = DetailActivity.class.getSimpleName();
@@ -65,141 +72,141 @@ public class DetailActivity extends AppCompatActivity
     private boolean mIsInQueue;
 
     @Override
-    protected void onCreate( Bundle savedInstanceState ) {
-        super.onCreate( savedInstanceState );
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
 
         try {
-            setContentView( R.layout.activity_detail );
-            getSupportActionBar().setDisplayHomeAsUpEnabled( true );
+            setContentView(R.layout.activity_detail);
+            Objects.requireNonNull(getSupportActionBar()).setDisplayHomeAsUpEnabled(true);
 
             long id;
             int quantity;
 
-            if ( savedInstanceState != null ) {
-                id = savedInstanceState.getLong( DatabaseContract.EngramEntry._ID );
-                quantity = savedInstanceState.getInt( DatabaseContract.QueueEntry.COLUMN_QUANTITY );
+            if (savedInstanceState != null) {
+                id = savedInstanceState.getLong(DatabaseContract.EngramEntry._ID);
+                quantity = savedInstanceState.getInt(DatabaseContract.QueueEntry.COLUMN_QUANTITY);
             } else {
-                id = getIntent().getExtras().getLong( REQUEST_ID );
+                id = Objects.requireNonNull(getIntent().getExtras()).getLong(REQUEST_ID);
                 quantity = 0;
             }
 
-            setShowcase( new Showcase( getApplicationContext(), id, quantity ) );
+            setShowcase(new Showcase(getApplicationContext(), id, quantity));
 
-            ImageView imageView = ( ImageView ) findViewById( R.id.imageView );
-            imageView.setContentDescription( "Image of " + getShowcase().getName() );
+            ImageView imageView = findViewById(R.id.imageView);
+            imageView.setContentDescription("Image of " + getShowcase().getName());
 
             String imagePath = "file:///android_asset/" + getShowcase().getImagePath();
-            Picasso.with( getApplicationContext() ).load( imagePath ).into( imageView );
+            Picasso.with(getApplicationContext()).load(imagePath).into(imageView);
 
-            TextView nameText = ( TextView ) findViewById( R.id.textView_name );
-            nameText.setText( getShowcase().getName() );
+            TextView nameText = findViewById(R.id.textView_name);
+            nameText.setText(getShowcase().getName());
 
-            Button decreaseButtonBy10 = ( Button ) findViewById( R.id.decreaseButtonBy10 );
-            decreaseButtonBy10.setOnClickListener( new View.OnClickListener() {
+            Button decreaseButtonBy10 = findViewById(R.id.decreaseButtonBy10);
+            decreaseButtonBy10.setOnClickListener(new View.OnClickListener() {
                 @Override
-                public void onClick( View v ) {
-                    decreaseQuantity( DECREMENT );
+                public void onClick(View v) {
+                    decreaseQuantity(DECREMENT);
                 }
-            } );
+            });
 
-            Button decreaseButton = ( Button ) findViewById( R.id.decreaseButton );
-            decreaseButton.setOnClickListener( new View.OnClickListener() {
+            Button decreaseButton = findViewById(R.id.decreaseButton);
+            decreaseButton.setOnClickListener(new View.OnClickListener() {
                 @Override
-                public void onClick( View v ) {
+                public void onClick(View v) {
                     decreaseQuantity();
                 }
-            } );
+            });
 
-            Button increaseButton = ( Button ) findViewById( R.id.increaseButton );
-            increaseButton.setOnClickListener( new View.OnClickListener() {
+            Button increaseButton = findViewById(R.id.increaseButton);
+            increaseButton.setOnClickListener(new View.OnClickListener() {
                 @Override
-                public void onClick( View v ) {
+                public void onClick(View v) {
                     increaseQuantity();
                 }
-            } );
+            });
 
-            Button increaseButtonBy10 = ( Button ) findViewById( R.id.increaseButtonBy10 );
-            increaseButtonBy10.setOnClickListener( new View.OnClickListener() {
+            Button increaseButtonBy10 = findViewById(R.id.increaseButtonBy10);
+            increaseButtonBy10.setOnClickListener(new View.OnClickListener() {
                 @Override
-                public void onClick( View v ) {
-                    increaseQuantity( INCREMENT );
+                public void onClick(View v) {
+                    increaseQuantity(INCREMENT);
                 }
-            } );
+            });
 
-            Button saveButton = ( Button ) findViewById( R.id.saveButton );
-            Button removeButton = ( Button ) findViewById( R.id.removeButton );
+            Button saveButton = findViewById(R.id.saveButton);
+            Button removeButton = findViewById(R.id.removeButton);
 
-            mIsInQueue = ( getShowcase().getQuantity() > 0 );
+            mIsInQueue = (getShowcase().getQuantity() > 0);
 
-            if ( mIsInQueue ) {
-                saveButton.setText( R.string.detail_content_save_button_update_queue );
+            if (mIsInQueue) {
+                saveButton.setText(R.string.detail_content_save_button_update_queue);
             } else {
-                removeButton.setEnabled( false );
-                saveButton.setText( R.string.detail_content_save_button_add_to_queue );
+                removeButton.setEnabled(false);
+                saveButton.setText(R.string.detail_content_save_button_add_to_queue);
             }
 
-            TextView categoryText = ( TextView ) findViewById( R.id.textView_navigation );
-            categoryText.setText( getShowcase().getCategoryHierarchy( getApplicationContext() ) );
+            TextView categoryText = findViewById(R.id.textView_navigation);
+            categoryText.setText(getShowcase().getCategoryHierarchy(getApplicationContext()));
 
             String craftedIn;
-            if ( getShowcase().getShowcaseEntry().getStations().size() == 1
-                    && getShowcase().getShowcaseEntry().getStations().keyAt( 0 ) == 0 )
-                craftedIn = getString( R.string.content_detail_crafted_in_no_station );
+            if (getShowcase().getShowcaseEntry().getStations().size() == 1
+                    && getShowcase().getShowcaseEntry().getStations().keyAt(0) == 0)
+                craftedIn = getString(R.string.content_detail_crafted_in_no_station);
             else
-                craftedIn = String.format( getString( R.string.content_detail_crafted_in_format ),
-                        getShowcase().getStationNameArrayAsText() );
+                craftedIn = String.format(getString(R.string.content_detail_crafted_in_format),
+                        getShowcase().getStationNameArrayAsText());
 
             String level;
-            if ( getShowcase().getRequiredLevel() == 1 )
-                level = getString( R.string.content_detail_required_level_any_level );
+            if (getShowcase().getRequiredLevel() == 1)
+                level = getString(R.string.content_detail_required_level_any_level);
             else
-                level = String.format( getString( R.string.content_detail_required_level_format ),
-                        getShowcase().getRequiredLevel() );
+                level = String.format(getString(R.string.content_detail_required_level_format),
+                        getShowcase().getRequiredLevel());
 
-            TextView descriptionText = ( TextView ) findViewById( R.id.textView_description );
-            descriptionText.setText( getShowcase().getDescription() );
+            TextView descriptionText = findViewById(R.id.textView_description);
+            descriptionText.setText(getShowcase().getDescription());
 
-            TextView stationText = ( TextView ) findViewById( R.id.textView_crafted_in );
-            stationText.setText( String.format( getString( R.string.content_detail_station_and_level_format ), craftedIn, level ) );
+            TextView stationText = findViewById(R.id.textView_crafted_in);
+            stationText.setText(String.format(getString(R.string.content_detail_station_and_level_format), craftedIn, level));
 
-            TextView dlcText = ( TextView ) findViewById( R.id.textView_dlc_name );
-            dlcText.setText( getShowcase().getDLCName() );
+            TextView dlcText = findViewById(R.id.textView_dlc_name);
+            dlcText.setText(getShowcase().getDLCName());
 
             updateQuantityText();
 
-            setAdapter( new ShowcaseResourceListAdapter( getShowcase() ) );
+            setAdapter(new ShowcaseResourceListAdapter(getShowcase()));
 
-            RecyclerView recyclerView = ( RecyclerView ) findViewById( R.id.resourceList );
-            recyclerView.setLayoutManager( new LinearLayoutManager( this ) );
-            recyclerView.setAdapter( getAdapter() );
+            RecyclerView recyclerView = findViewById(R.id.resourceList);
+            recyclerView.setLayoutManager(new LinearLayoutManager(this));
+            recyclerView.setAdapter(getAdapter());
 
-            mAdUtil = new AdUtil( this, R.id.content_detail );
-        } catch ( Exception e ) {
-            FinishActivityWithError( e );
+            mAdUtil = new AdUtil(this, R.id.content_detail);
+        } catch (Exception e) {
+            FinishActivityWithError(e);
         }
 
     }
 
     @Override
-    public boolean onCreateOptionsMenu( Menu menu ) {
-        getMenuInflater().inflate( R.menu.menu_detail, menu );
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.menu_detail, menu);
         return true;
     }
 
     @Override
-    public boolean onOptionsItemSelected( MenuItem item ) {
+    public boolean onOptionsItemSelected(MenuItem item) {
 
-        switch ( item.getItemId() ) {
+        switch (item.getItemId()) {
             case android.R.id.home:
                 FinishActivityWithoutResult();
                 return true;
 
             case R.id.action_feedback_correction:
-                FeedbackUtil.composeCorrectionEmail( DetailActivity.this, getShowcase().getName() );
+                FeedbackUtil.composeCorrectionEmail(DetailActivity.this, getShowcase().getName());
                 break;
         }
 
-        return super.onOptionsItemSelected( item );
+        return super.onOptionsItemSelected(item);
     }
 
     @Override
@@ -219,7 +226,7 @@ public class DetailActivity extends AppCompatActivity
     protected void onPause() {
         super.onPause();
 
-        CraftingQueue.getInstance().pause( this );
+        CraftingQueue.getInstance().pause(this);
         mAdUtil.pause();
     }
 
@@ -232,14 +239,14 @@ public class DetailActivity extends AppCompatActivity
     }
 
     @Override
-    protected void onSaveInstanceState( Bundle outState ) {
-        outState.putLong( DatabaseContract.EngramEntry._ID, getShowcase().getId() );
-        outState.putInt( DatabaseContract.QueueEntry.COLUMN_QUANTITY, getShowcase().getQuantity() );
+    protected void onSaveInstanceState(Bundle outState) {
+        outState.putLong(DatabaseContract.EngramEntry._ID, getShowcase().getId());
+        outState.putInt(DatabaseContract.QueueEntry.COLUMN_QUANTITY, getShowcase().getQuantity());
 
-        super.onSaveInstanceState( outState );
+        super.onSaveInstanceState(outState);
     }
 
-    private void setShowcase( Showcase showcase ) {
+    private void setShowcase(Showcase showcase) {
         this.mShowcase = showcase;
     }
 
@@ -247,7 +254,7 @@ public class DetailActivity extends AppCompatActivity
         return mShowcase;
     }
 
-    private void setAdapter( ShowcaseResourceListAdapter adapter ) {
+    private void setAdapter(ShowcaseResourceListAdapter adapter) {
         this.mAdapter = adapter;
     }
 
@@ -260,8 +267,8 @@ public class DetailActivity extends AppCompatActivity
         notifyAndUpdate();
     }
 
-    public void increaseQuantity( int increment ) {
-        getShowcase().increaseQuantity( increment );
+    public void increaseQuantity(int increment) {
+        getShowcase().increaseQuantity(increment);
         notifyAndUpdate();
     }
 
@@ -270,8 +277,8 @@ public class DetailActivity extends AppCompatActivity
         notifyAndUpdate();
     }
 
-    public void decreaseQuantity( int decrement ) {
-        getShowcase().decreaseQuantity( decrement );
+    public void decreaseQuantity(int decrement) {
+        getShowcase().decreaseQuantity(decrement);
         notifyAndUpdate();
     }
 
@@ -286,73 +293,73 @@ public class DetailActivity extends AppCompatActivity
     }
 
     private void updateQuantityText() {
-        TextView quantityEditText = ( TextView ) findViewById( R.id.quantityEditText );
-        quantityEditText.setText( getShowcase().getQuantityText() );
+        TextView quantityEditText = findViewById(R.id.quantityEditText);
+        quantityEditText.setText(getShowcase().getQuantityText());
     }
 
-    public void RemoveAndFinish( View view ) {
-        FinishActivityWithResult( REMOVE );
+    public void RemoveAndFinish(View view) {
+        FinishActivityWithResult(REMOVE);
     }
 
-    public void SaveAndFinish( View view ) {
-        if ( mIsInQueue )
-            FinishActivityWithResult( UPDATE );
+    public void SaveAndFinish(View view) {
+        if (mIsInQueue)
+            FinishActivityWithResult(UPDATE);
         else
-            FinishActivityWithResult( ADD );
+            FinishActivityWithResult(ADD);
     }
 
-    private void FinishActivityWithError( Exception e ) {
+    private void FinishActivityWithError(Exception e) {
         Intent returnIntent = getIntent();
 
-        returnIntent.putExtra( RESULT_CODE, ERROR );
-        returnIntent.putExtra( RESULT_EXTRA_NAME, e );
+        returnIntent.putExtra(RESULT_CODE, ERROR);
+        returnIntent.putExtra(RESULT_EXTRA_NAME, e);
 
-        setResult( RESULT_CANCELED, returnIntent );
+        setResult(RESULT_CANCELED, returnIntent);
 
         finish();
     }
 
-    private void FinishActivityWithResult( int resultCode ) {
+    private void FinishActivityWithResult(int resultCode) {
         Intent returnIntent = getIntent();
 
-        returnIntent.putExtra( RESULT_CODE, resultCode );
-        returnIntent.putExtra( RESULT_EXTRA_NAME, getShowcase().getName() );
-        returnIntent.putExtra( RESULT_EXTRA_ID, getShowcase().getId() );
+        returnIntent.putExtra(RESULT_CODE, resultCode);
+        returnIntent.putExtra(RESULT_EXTRA_NAME, getShowcase().getName());
+        returnIntent.putExtra(RESULT_EXTRA_ID, getShowcase().getId());
 
-        switch ( resultCode ) {
+        switch (resultCode) {
             case REMOVE:
-                CraftingQueue.getInstance().delete( getShowcase().getId() );
+                CraftingQueue.getInstance().delete(getShowcase().getId());
                 break;
 
             case ADD:
             case UPDATE:
                 int quantity = getShowcase().getQuantity();
 
-                if ( quantity == 0 && !mIsInQueue )
+                if (quantity == 0 && !mIsInQueue)
                     quantity++;
 
-                CraftingQueue.getInstance().setQuantity( this, getShowcase().getId(), quantity );
+                CraftingQueue.getInstance().setQuantity(this, getShowcase().getId(), quantity);
 
-                returnIntent.putExtra( RESULT_EXTRA_QUANTITY, quantity );
+                returnIntent.putExtra(RESULT_EXTRA_QUANTITY, quantity);
                 break;
         }
 
-        setResult( RESULT_OK, returnIntent );
+        setResult(RESULT_OK, returnIntent);
         finish();
     }
 
     private void FinishActivityWithoutResult() {
-        setResult( RESULT_CANCELED, getIntent() );
+        setResult(RESULT_CANCELED, getIntent());
         finish();
     }
 
     @Override
-    public void onException( String tag, Exception e ) {
-        ExceptionUtil.SendErrorReport( tag, e );
+    public void onException(String tag, Exception e) {
+        ExceptionUtil.SendErrorReport(tag, e);
     }
 
     @Override
-    public void onFatalException( String tag, Exception e ) {
-        ExceptionUtil.SendErrorReportWithAlertDialog( DetailActivity.this, tag, e );
+    public void onFatalException(String tag, Exception e) {
+        ExceptionUtil.SendErrorReportWithAlertDialog(DetailActivity.this, tag, e);
     }
 }
