@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.listeners;
 
 import java.util.ArrayList;
@@ -13,13 +29,13 @@ public class ExceptionObserver {
     private List<Listener> mListeners;
 
     public interface Listener {
-        void onException( String tag, Exception e );
+        void onException(String tag, Exception e);
 
-        void onFatalException( String tag, Exception e );
+        void onFatalException(String tag, Exception e);
     }
 
     public static ExceptionObserver getInstance() {
-        if ( sInstance == null )
+        if (sInstance == null)
             sInstance = new ExceptionObserver();
 
         return sInstance;
@@ -29,40 +45,40 @@ public class ExceptionObserver {
         mListeners = new ArrayList<>();
     }
 
-    public void registerListener( Listener listener ) {
-        if ( !mListeners.contains( listener ) )
-            mListeners.add( listener );
+    public void registerListener(Listener listener) {
+        if (!mListeners.contains(listener))
+            mListeners.add(listener);
     }
 
-    public void unregisterListener( Listener listener ) {
-        mListeners.remove( listener );
+    public void unregisterListener(Listener listener) {
+        mListeners.remove(listener);
     }
 
-    public void notifyExceptionCaught( String tag, Exception e ) {
-        if ( mListeners.size() > 0 ) {
-            for ( Listener listener : mListeners ) {
-                listener.onException( tag, e );
+    public void notifyExceptionCaught(String tag, Exception e) {
+        if (mListeners.size() > 0) {
+            for (Listener listener : mListeners) {
+                listener.onException(tag, e);
             }
         } else {
-            ExceptionUtil.SendErrorReport( tag, e );
+            ExceptionUtil.SendErrorReport(tag, e);
         }
     }
 
-    public void notifyFatalExceptionCaught( String tag, Exception e ) {
-        if ( mListeners.size() > 0 ) {
-            for ( Listener listener : mListeners ) {
-                listener.onFatalException( tag, e );
+    public void notifyFatalExceptionCaught(String tag, Exception e) {
+        if (mListeners.size() > 0) {
+            for (Listener listener : mListeners) {
+                listener.onFatalException(tag, e);
             }
         } else {
-            ExceptionUtil.SendErrorReport( tag, e );
+            ExceptionUtil.SendErrorReport(tag, e);
         }
     }
 
     public void throwException() {
-        notifyExceptionCaught( TAG, new Exception( "Manually thrown exception." ) );
+        notifyExceptionCaught(TAG, new Exception("Manually thrown exception."));
     }
 
     public void throwFatalException() {
-        notifyFatalExceptionCaught( TAG, new Exception( "Manually thrown exception." ) );
+        notifyFatalExceptionCaught(TAG, new Exception("Manually thrown exception."));
     }
 }
