@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.model;
 
 import android.content.Context;
@@ -19,18 +35,6 @@ import arc.resource.calculator.model.engram.QueueEngram;
 import arc.resource.calculator.util.ExceptionUtil;
 import arc.resource.calculator.util.PrefsUtil;
 
-/**
- * Copyright (C) 2016, Jared Stone
- * -
- * Author: Jared Stone
- * Title: A:RC, a resource calculator for ARK:Survival Evolved
- * -
- * Web: https://github.com/jaredstone1982/CraftingCalcARK
- * Email: jaredstone1982@gmail.com
- * Twitter: @MasterxOfxNone
- * -
- * This work is licensed under a Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International License.
- */
 public class CraftingQueue {
     private static final String TAG = CraftingQueue.class.getSimpleName();
 
@@ -39,22 +43,22 @@ public class CraftingQueue {
     private static QueueAdapter.Observer mViewObserver;
 
     private CraftingQueue() {
-        Log.d( TAG, "CraftingQueue: new CraftingQueue" );
+        Log.d(TAG, "CraftingQueue: new CraftingQueue");
 
-        setQueue( new QueueMap() );
+        setQueue(new QueueMap());
 
-        PrefsObserver.getInstance().registerListener( TAG, new PrefsObserver.Listener() {
+        PrefsObserver.getInstance().registerListener(TAG, new PrefsObserver.Listener() {
             @Override
-            public void onPreferencesChanged( boolean dlcValueChange, boolean categoryPrefChange, boolean stationPrefChange,
-                                              boolean levelPrefChange, boolean levelValueChange, boolean refinedPrefChange ) {
-                if ( dlcValueChange )
+            public void onPreferencesChanged(boolean dlcValueChange, boolean categoryPrefChange, boolean stationPrefChange,
+                                             boolean levelPrefChange, boolean levelValueChange, boolean refinedPrefChange) {
+                if (dlcValueChange)
                     clearQueue();
             }
-        } );
+        });
     }
 
     public static CraftingQueue getInstance() {
-        if ( sInstance == null )
+        if (sInstance == null)
             sInstance = new CraftingQueue();
 
         return sInstance;
@@ -63,19 +67,19 @@ public class CraftingQueue {
     public void resume() {
     }
 
-    public void pause( Context context ) {
-        saveQueueToPrefs( context );
+    public void pause(Context context) {
+        saveQueueToPrefs(context);
     }
 
     public void destroy() {
-        PrefsObserver.getInstance().unregisterListener( TAG );
+        PrefsObserver.getInstance().unregisterListener(TAG);
     }
 
     private QueueMap getQueue() {
         return mQueue;
     }
 
-    private void setQueue( QueueMap queue ) {
+    private void setQueue(QueueMap queue) {
         this.mQueue = queue;
     }
 
@@ -83,7 +87,7 @@ public class CraftingQueue {
         return mViewObserver;
     }
 
-    public void setObserver( QueueAdapter.Observer observer ) {
+    public void setObserver(QueueAdapter.Observer observer) {
         mViewObserver = observer;
     }
 
@@ -91,125 +95,125 @@ public class CraftingQueue {
         return getQueue().size();
     }
 
-    public boolean contains( long id ) {
-        return getQueue().contains( id );
+    public boolean contains(long id) {
+        return getQueue().contains(id);
     }
 
     public boolean isQueueEmpty() {
         return getSize() == 0;
     }
 
-    public QueueEngram getCraftable( int position ) {
-        return getQueue().valueAt( position );
+    public QueueEngram getCraftable(int position) {
+        return getQueue().valueAt(position);
     }
 
-    public QueueEngram getCraftable( long id ) {
-        return getQueue().get( id );
+    public QueueEngram getCraftable(long id) {
+        return getQueue().get(id);
     }
 
-    private void updateCraftable( int position, QueueEngram craftable ) {
-        getQueue().setValueAt( position, craftable );
+    private void updateCraftable(int position, QueueEngram craftable) {
+        getQueue().setValueAt(position, craftable);
     }
 
-    private void addCraftable( QueueEngram craftable ) {
-        getQueue().add( craftable.getId(), craftable );
+    private void addCraftable(QueueEngram craftable) {
+        getQueue().add(craftable.getId(), craftable);
     }
 
-    private void removeCraftable( int position ) {
-        getQueue().removeAt( position );
+    private void removeCraftable(int position) {
+        getQueue().removeAt(position);
     }
 
     private void removeAllCraftables() {
         getQueue().clear();
     }
 
-    private int getPosition( long id ) {
-        return getQueue().indexOfKey( id );
+    private int getPosition(long id) {
+        return getQueue().indexOfKey(id);
     }
 
-    public void increaseQuantity( int position ) {
-        QueueEngram craftable = getCraftable( position );
+    public void increaseQuantity(int position) {
+        QueueEngram craftable = getCraftable(position);
 
         craftable.increaseQuantity();
 
         // update item in list
-        updateCraftable( position, craftable );
+        updateCraftable(position, craftable);
 
         // notify list adapter of changes
-        if ( getObserver() != null )
-            getObserver().notifyItemChanged( position );
+        if (getObserver() != null)
+            getObserver().notifyItemChanged(position);
 
         // notify outside listeners of changes
-        QueueObserver.getInstance().notifyItemChanged( craftable.getId(), craftable.getQuantity() );
+        QueueObserver.getInstance().notifyItemChanged(craftable.getId(), craftable.getQuantity());
     }
 
-    public void increaseQuantity( QueueEngram craftable ) {
+    public void increaseQuantity(QueueEngram craftable) {
         long id = craftable.getId();
 
         craftable.increaseQuantity();
 
-        if ( getQueue().contains( id ) ) {
-            int position = getPosition( id );
+        if (getQueue().contains(id)) {
+            int position = getPosition(id);
 
             // update item in list
-            updateCraftable( position, craftable );
+            updateCraftable(position, craftable);
 
             // notify list adapter of changes
-            if ( getObserver() != null )
-                getObserver().notifyItemChanged( position );
+            if (getObserver() != null)
+                getObserver().notifyItemChanged(position);
 
             // notify outside listeners of changes
-            QueueObserver.getInstance().notifyItemChanged( id, craftable.getQuantity() );
+            QueueObserver.getInstance().notifyItemChanged(id, craftable.getQuantity());
         } else {
-            insert( craftable );
+            insert(craftable);
         }
     }
 
-    public void setQuantity( Context context, long id, int quantity ) {
-        if ( getQueue().contains( id ) ) {
-            int position = getPosition( id );
-            if ( quantity > 0 ) {
+    public void setQuantity(Context context, long id, int quantity) {
+        if (getQueue().contains(id)) {
+            int position = getPosition(id);
+            if (quantity > 0) {
                 // get object from list
-                QueueEngram craftable = getCraftable( position );
+                QueueEngram craftable = getCraftable(position);
 
                 // update object's quantity
-                craftable.setQuantity( quantity );
+                craftable.setQuantity(quantity);
 
                 // update item in list
-                updateCraftable( position, craftable );
+                updateCraftable(position, craftable);
 
                 // notify list adapter of changes
-                if ( getObserver() != null )
-                    getObserver().notifyItemChanged( position );
+                if (getObserver() != null)
+                    getObserver().notifyItemChanged(position);
 
                 // notify outside listeners of changes
-                QueueObserver.getInstance().notifyItemChanged( id, craftable.getQuantity() );
+                QueueObserver.getInstance().notifyItemChanged(id, craftable.getQuantity());
             } else {
-                delete( position );
+                delete(position);
             }
         } else {
-            if ( quantity > 0 ) {
-                insert( context, id, quantity );
+            if (quantity > 0) {
+                insert(context, id, quantity);
             }
         }
     }
 
-    private void insert( Context context, long id, int quantity ) {
-        Log.d( TAG, "insert: " );
+    private void insert(Context context, long id, int quantity) {
+        Log.d(TAG, "insert: ");
         try {
-            insert( QueryForCraftable( context, id, quantity ) );
-        } catch ( Exception e ) {
+            insert(QueryForCraftable(context, id, quantity));
+        } catch (Exception e) {
             // notify ViewSwitcher of error
-            Log.e( TAG, "insert: ", e );
-            if ( getObserver() != null )
-                getObserver().notifyExceptionCaught( e );
+            Log.e(TAG, "insert: ", e);
+            if (getObserver() != null)
+                getObserver().notifyExceptionCaught(e);
         }
     }
 
-    private void insert( QueueEngram craftable ) {
-        Log.d( TAG, "insert: " + craftable.toString() );
+    private void insert(QueueEngram craftable) {
+        Log.d(TAG, "insert: " + craftable.toString());
         // add craftable to list
-        addCraftable( craftable );
+        addCraftable(craftable);
 
         // sort list by name
         getQueue().sort();
@@ -219,29 +223,29 @@ public class CraftingQueue {
 
         // notify list adapter of changes
 //        getObserver().notifyItemInserted( position );
-        if ( getObserver() != null )
+        if (getObserver() != null)
             getObserver().notifyDataSetPopulated();
 
         // notify outside listeners of changes
-        QueueObserver.getInstance().notifyItemChanged( craftable.getId(), craftable.getQuantity() );
+        QueueObserver.getInstance().notifyItemChanged(craftable.getId(), craftable.getQuantity());
     }
 
-    private void delete( int position ) {
-        long id = getCraftable( position ).getId();
+    private void delete(int position) {
+        long id = getCraftable(position).getId();
 
         // remove craftable from list
-        removeCraftable( position );
+        removeCraftable(position);
 
-        if ( getQueue().size() > 0 ) {
+        if (getQueue().size() > 0) {
             // notify list adapter of changes
-            if ( getObserver() != null )
-                getObserver().notifyItemRemoved( position );
+            if (getObserver() != null)
+                getObserver().notifyItemRemoved(position);
 
             // notify outside listeners of changes
-            QueueObserver.getInstance().notifyItemRemoved( id );
+            QueueObserver.getInstance().notifyItemRemoved(id);
         } else {
             // notify ViewSwitcher of empty status
-            if ( getObserver() != null )
+            if (getObserver() != null)
                 getObserver().notifyDataSetEmpty();
 
             // notify outside listeners of changes
@@ -249,9 +253,9 @@ public class CraftingQueue {
         }
     }
 
-    public void delete( long id ) {
-        if ( getQueue().contains( id ) ) {
-            delete( getPosition( id ) );
+    public void delete(long id) {
+        if (getQueue().contains(id)) {
+            delete(getPosition(id));
         }
     }
 
@@ -260,87 +264,87 @@ public class CraftingQueue {
         removeAllCraftables();
 
         // notify ViewSwitcher of empty status
-        if ( getObserver() != null )
+        if (getObserver() != null)
             getObserver().notifyDataSetEmpty();
 
         // notify outside listeners of changes
         QueueObserver.getInstance().notifyDataSetEmpty();
     }
 
-    private void saveQueueToPrefs( Context context ) {
-        PrefsUtil.getInstance( context ).saveCraftingQueueJSONString( convertQueueToJSONString() );
+    private void saveQueueToPrefs(Context context) {
+        PrefsUtil.getInstance(context).saveCraftingQueueJSONString(convertQueueToJSONString());
     }
 
     private String convertQueueToJSONString() {
         try {
             JSONArray json = new JSONArray();
 
-            for ( int i = 0; i < getQueue().size(); i++ ) {
-                QueueEngram craftable = getCraftable( i );
+            for (int i = 0; i < getQueue().size(); i++) {
+                QueueEngram craftable = getCraftable(i);
 
                 JSONObject object = new JSONObject();
-                object.put( DatabaseContract.EngramEntry._ID, craftable.getId() );
-                object.put( DatabaseContract.QueueEntry.COLUMN_QUANTITY, craftable.getQuantity() );
+                object.put(DatabaseContract.EngramEntry._ID, craftable.getId());
+                object.put(DatabaseContract.QueueEntry.COLUMN_QUANTITY, craftable.getQuantity());
 
-                json.put( object );
+                json.put(object);
             }
 
             return json.toString();
-        } catch ( JSONException e ) {
+        } catch (JSONException e) {
             return null;
         }
     }
 
-    private LongSparseArray<Integer> convertJSONStringToQueue( Context context ) {
+    private LongSparseArray<Integer> convertJSONStringToQueue(Context context) {
         try {
-            String jsonString = PrefsUtil.getInstance( context ).getCraftingQueueJSONString();
+            String jsonString = PrefsUtil.getInstance(context).getCraftingQueueJSONString();
 
-            if ( jsonString == null )
-                return new LongSparseArray<>( 0 );
+            if (jsonString == null)
+                return new LongSparseArray<>(0);
 
-            JSONArray json = new JSONArray( jsonString );
+            JSONArray json = new JSONArray(jsonString);
 
-            LongSparseArray<Integer> convertedQueueArray = new LongSparseArray<>( json.length() );
-            for ( int i = 0; i < json.length(); i++ ) {
-                JSONObject object = json.getJSONObject( i );
+            LongSparseArray<Integer> convertedQueueArray = new LongSparseArray<>(json.length());
+            for (int i = 0; i < json.length(); i++) {
+                JSONObject object = json.getJSONObject(i);
 
-                long engramId = object.getLong( DatabaseContract.EngramEntry._ID );
-                int quantity = object.getInt( DatabaseContract.QueueEntry.COLUMN_QUANTITY );
+                long engramId = object.getLong(DatabaseContract.EngramEntry._ID);
+                int quantity = object.getInt(DatabaseContract.QueueEntry.COLUMN_QUANTITY);
 
-                convertedQueueArray.put( engramId, quantity );
+                convertedQueueArray.put(engramId, quantity);
             }
 
             return convertedQueueArray;
-        } catch ( Exception e ) {
-            return new LongSparseArray<>( 0 );
+        } catch (Exception e) {
+            return new LongSparseArray<>(0);
         }
     }
 
-    private QueueEngram QueryForCraftable( Context context, long engramId, int quantity ) throws Exception {
-        long dlc_id = PrefsUtil.getInstance( context ).getDLCPreference();
-        Uri uri = DatabaseContract.EngramEntry.buildUriWithId( dlc_id, engramId );
+    private QueueEngram QueryForCraftable(Context context, long engramId, int quantity) throws Exception {
+        long dlc_id = PrefsUtil.getInstance(context).getDLCPreference();
+        Uri uri = DatabaseContract.EngramEntry.buildUriWithId(dlc_id, engramId);
 
-        try ( Cursor cursor = context.getContentResolver().query( uri, null, null, null, null ) ) {
+        try (Cursor cursor = context.getContentResolver().query(uri, null, null, null, null)) {
 
-            if ( cursor == null )
-                throw new ExceptionUtil.CursorNullException( uri );
+            if (cursor == null)
+                throw new ExceptionUtil.CursorNullException(uri);
 
-            if ( !cursor.moveToFirst() )
-                throw new ExceptionUtil.CursorEmptyException( uri );
+            if (!cursor.moveToFirst())
+                throw new ExceptionUtil.CursorEmptyException(uri);
 
             return new QueueEngram(
                     engramId,
-                    cursor.getString( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_NAME ) ),
-                    cursor.getString( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_IMAGE_FOLDER ) ),
-                    cursor.getString( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_IMAGE_FILE ) ),
-                    cursor.getInt( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_YIELD ) ),
-                    quantity );
+                    cursor.getString(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_NAME)),
+                    cursor.getString(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_IMAGE_FOLDER)),
+                    cursor.getString(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_IMAGE_FILE)),
+                    cursor.getInt(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_YIELD)),
+                    quantity);
         }
     }
 
-    public void fetch( Context context ) {
-        Log.d( TAG, "fetch: " );
-        new FetchDataTask( context ).execute();
+    public void fetch(Context context) {
+        Log.d(TAG, "fetch: ");
+        new FetchDataTask(context).execute();
     }
 
     private class FetchDataTask extends AsyncTask<Void, Void, Boolean> {
@@ -351,60 +355,60 @@ public class CraftingQueue {
 
         private Exception mException;
 
-        FetchDataTask( Context context ) {
+        FetchDataTask(Context context) {
             this.mContext = context.getApplicationContext();
             this.mTempQueueMap = new QueueMap();
         }
 
         @Override
         protected void onPreExecute() {
-            if ( getObserver() != null )
+            if (getObserver() != null)
                 getObserver().notifyInitializing();
         }
 
         @Override
-        protected void onPostExecute( Boolean querySuccessful ) {
-            Log.d( TAG, "onPostExecute: " + querySuccessful );
-            if ( querySuccessful ) {
-                if ( mTempQueueMap.size() > 0 ) {
-                    setQueue( mTempQueueMap );
+        protected void onPostExecute(Boolean querySuccessful) {
+            Log.d(TAG, "onPostExecute: " + querySuccessful);
+            if (querySuccessful) {
+                if (mTempQueueMap.size() > 0) {
+                    setQueue(mTempQueueMap);
 
-                    if ( getObserver() != null )
+                    if (getObserver() != null)
                         getObserver().notifyDataSetPopulated();
 
                     QueueObserver.getInstance().notifyDataSetPopulated();
                 } else {
-                    if ( getObserver() != null )
+                    if (getObserver() != null)
                         getObserver().notifyDataSetEmpty();
                 }
             } else {
-                if ( mException == null )
-                    mException = new Exception( "Nullified Exception caught." );
+                if (mException == null)
+                    mException = new Exception("Nullified Exception caught.");
 
-                if ( getObserver() != null )
-                    getObserver().notifyExceptionCaught( mException );
+                if (getObserver() != null)
+                    getObserver().notifyExceptionCaught(mException);
             }
         }
 
         @Override
-        protected Boolean doInBackground( Void... params ) {
+        protected Boolean doInBackground(Void... params) {
 
             try {
-                LongSparseArray<Integer> savedQueue = convertJSONStringToQueue( mContext );
+                LongSparseArray<Integer> savedQueue = convertJSONStringToQueue(mContext);
 
-                if ( savedQueue.size() > 0 ) {
-                    for ( int i = 0; i < savedQueue.size(); i++ ) {
-                        long id = savedQueue.keyAt( i );
-                        int quantity = savedQueue.valueAt( i );
+                if (savedQueue.size() > 0) {
+                    for (int i = 0; i < savedQueue.size(); i++) {
+                        long id = savedQueue.keyAt(i);
+                        int quantity = savedQueue.valueAt(i);
 
-                        QueueEngram craftable = QueryForCraftable( mContext, id, quantity );
+                        QueueEngram craftable = QueryForCraftable(mContext, id, quantity);
 
-                        mTempQueueMap.add( id, craftable );
+                        mTempQueueMap.add(id, craftable);
                     }
                 }
 
                 return true;
-            } catch ( Exception e ) {
+            } catch (Exception e) {
                 mException = e;
                 return false;
             }
@@ -418,18 +422,18 @@ public class CraftingQueue {
         }
 
         @Override
-        public QueueEngram get( long key ) {
-            return ( QueueEngram ) super.get( key );
+        public QueueEngram get(long key) {
+            return (QueueEngram) super.get(key);
         }
 
         @Override
-        public QueueEngram valueAt( int position ) {
-            return ( QueueEngram ) super.valueAt( position );
+        public QueueEngram valueAt(int position) {
+            return (QueueEngram) super.valueAt(position);
         }
 
         @Override
-        public Comparable getComparable( int position ) {
-            return valueAt( position ).getName();
+        public Comparable getComparable(int position) {
+            return valueAt(position).getName();
         }
     }
 }
