@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.model;
 
 import android.util.Base64;
@@ -6,20 +22,20 @@ import javax.annotation.Nonnull;
 
 public class Encryption {
     @Nonnull
-    public static String decrypt( @Nonnull String message, @Nonnull String salt ) {
-        return xor( new String( Base64.decode( message, 0 ) ), salt );
+    public static String decrypt(@Nonnull String message, @Nonnull String salt) {
+        return xor(new String(Base64.decode(message, 0)), salt);
     }
 
     @Nonnull
-    public static String encrypt( @Nonnull String message, @Nonnull String salt ) {
-        return new String( Base64.encode( xor( message, salt ).getBytes(), 0 ) );
+    public static String encrypt(@Nonnull String message, @Nonnull String salt) {
+        return new String(Base64.encode(xor(message, salt).getBytes(), 0));
     }
 
     /**
      * Encrypts or decrypts a base-64 string using a XOR cipher.
      */
     @Nonnull
-    private static String xor( @Nonnull String message, @Nonnull String salt ) {
+    private static String xor(@Nonnull String message, @Nonnull String salt) {
         final char[] m = message.toCharArray();
         final int ml = m.length;
 
@@ -27,9 +43,9 @@ public class Encryption {
         final int sl = s.length;
 
         final char[] res = new char[ml];
-        for ( int i = 0; i < ml; i++ ) {
-            res[i] = ( char ) ( m[i] ^ s[i % sl] );
+        for (int i = 0; i < ml; i++) {
+            res[i] = (char) (m[i] ^ s[i % sl]);
         }
-        return new String( res );
+        return new String(res);
     }
 }
