@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.model;
 
 import android.content.Context;
@@ -19,18 +35,6 @@ import arc.resource.calculator.util.ExceptionUtil;
 import arc.resource.calculator.util.PrefsUtil;
 import arc.resource.calculator.util.Util;
 
-/**
- * Copyright (C) 2016, Jared Stone
- * -
- * Author: Jared Stone
- * Title: A:RC, a resource calculator for ARK:Survival Evolved
- * -
- * Web: https://github.com/jaredstone1982/CraftingCalcARK
- * Email: jaredstone1982@gmail.com
- * Twitter: @MasterxOfxNone
- * -
- * This work is licensed under a Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International License.
- */
 public class Showcase {
     private static final String TAG = Showcase.class.getSimpleName();
 
@@ -38,17 +42,17 @@ public class Showcase {
 
     private ShowcaseEntry mShowcaseEntry;
 
-    public Showcase( Context context, long _id, int quantity )
+    public Showcase(Context context, long _id, int quantity)
             throws ExceptionUtil.CursorEmptyException, ExceptionUtil.CursorNullException {
-        long dlc_id = PrefsUtil.getInstance( context ).getDLCPreference();
+        long dlc_id = PrefsUtil.getInstance(context).getDLCPreference();
 
-        setContext( context );
-        setShowcaseEntry( QueryForDetails( context, dlc_id, _id, quantity ) );
+        setContext(context);
+        setShowcaseEntry(QueryForDetails(context, dlc_id, _id, quantity));
     }
 
-    private void setShowcaseEntry( ShowcaseEntry showcaseEntry ) {
-        if ( showcaseEntry == null )
-            Log.d( TAG, "showcaseEntry is null?" );
+    private void setShowcaseEntry(ShowcaseEntry showcaseEntry) {
+        if (showcaseEntry == null)
+            Log.d(TAG, "showcaseEntry is null?");
 
         mShowcaseEntry = showcaseEntry;
     }
@@ -57,7 +61,7 @@ public class Showcase {
         return mShowcaseEntry;
     }
 
-    private void setContext( Context context ) {
+    private void setContext(Context context) {
         this.mContext = context.getApplicationContext();
     }
 
@@ -77,8 +81,8 @@ public class Showcase {
         return getShowcaseEntry().getCraftable().getName();
     }
 
-    public QueueResource getResource( int position ) throws Exception {
-        return getShowcaseEntry().getResource( position );
+    public QueueResource getResource(int position) throws Exception {
+        return getShowcaseEntry().getResource(position);
     }
 
     public int getQuantity() {
@@ -86,7 +90,7 @@ public class Showcase {
     }
 
     public String getQuantityText() {
-        return String.valueOf( getShowcaseEntry().getCraftable().getQuantityWithYield() );
+        return String.valueOf(getShowcaseEntry().getCraftable().getQuantityWithYield());
     }
 
     public int getRequiredLevel() {
@@ -94,7 +98,7 @@ public class Showcase {
     }
 
     public String getDLCName() throws ExceptionUtil.CursorEmptyException, ExceptionUtil.CursorNullException {
-        return QueryForDLCName( getContext(), getShowcaseEntry().getDLCId() );
+        return QueryForDLCName(getContext(), getShowcaseEntry().getDLCId());
     }
 
     public String getDescription() {
@@ -102,50 +106,47 @@ public class Showcase {
     }
 
     public String getStationNameArrayAsText() {
-        Log.d( TAG, "getStationNameArrayAsText: " + getShowcaseEntry().getStations().toString() );
+        Log.d(TAG, "getStationNameArrayAsText: " + getShowcaseEntry().getStations().toString());
 
         int size = getShowcaseEntry().getStations().size();
 
-        if ( size == 1 )
-            return getShowcaseEntry().getStations().valueAt( 0 ).getName();
+        if (size == 1)
+            return getShowcaseEntry().getStations().valueAt(0).getName();
 
-        if ( size == 2 )
-            return String.format( getContext().getString( R.string.content_detail_crafted_in_double_format ),
-                    getShowcaseEntry().getStations().valueAt( 0 ).getName(),
-                    getShowcaseEntry().getStations().valueAt( 1 ).getName() );
+        if (size == 2)
+            return String.format(getContext().getString(R.string.content_detail_crafted_in_double_format),
+                    getShowcaseEntry().getStations().valueAt(0).getName(),
+                    getShowcaseEntry().getStations().valueAt(1).getName());
 
         StringBuilder builder = new StringBuilder();
-        for ( int i = 0; i < size; i++ ) {
-            if ( i > 0 ) {
-                if ( i == size - 1 )
-                    builder.append( String.format( getContext().getString( R.string.content_detail_crafted_in_multiple_last_format ),
-                            getShowcaseEntry().getStations().valueAt( i ).getName() ) );
+        for (int i = 0; i < size; i++) {
+            if (i > 0) {
+                if (i == size - 1)
+                    builder.append(String.format(getContext().getString(R.string.content_detail_crafted_in_multiple_last_format),
+                            getShowcaseEntry().getStations().valueAt(i).getName()));
                 else
-                    builder.append( String.format( getContext().getString( R.string.content_detail_crafted_in_multiple_format ),
-                            getShowcaseEntry().getStations().valueAt( i ).getName() ) );
+                    builder.append(String.format(getContext().getString(R.string.content_detail_crafted_in_multiple_format),
+                            getShowcaseEntry().getStations().valueAt(i).getName()));
             } else
-                builder.append( getShowcaseEntry().getStations().valueAt( i ).getName() );
+                builder.append(getShowcaseEntry().getStations().valueAt(i).getName());
         }
 
         return builder.toString();
     }
 
-    public String getCategoryHierarchy( Context context )
+    public String getCategoryHierarchy(Context context)
             throws ExceptionUtil.CursorEmptyException, ExceptionUtil.CursorNullException {
-        Category category = QueryForCategoryDetails( context, getShowcaseEntry().getDLCId(), getShowcaseEntry().getCraftable().getCategoryId() );
-
-        if ( category == null ) return null;
+        Category category = QueryForCategoryDetails(context, getShowcaseEntry().getDLCId(), getShowcaseEntry().getCraftable().getCategoryId());
 
         long parent_id = category.getParent();
 
-        StringBuilder builder = new StringBuilder( category.getName() );
-        while ( parent_id > 0 ) {
-            category = QueryForCategoryDetails( context, getShowcaseEntry().getDLCId(), parent_id );
-            if ( category == null ) break;
+        StringBuilder builder = new StringBuilder(category.getName());
+        while (parent_id > 0) {
+            category = QueryForCategoryDetails(context, getShowcaseEntry().getDLCId(), parent_id);
 
             parent_id = category.getParent();
 
-            builder.insert( 0, category.getName() + "/" );
+            builder.insert(0, category.getName() + "/");
         }
 
         return builder.toString();
@@ -155,166 +156,166 @@ public class Showcase {
         getShowcaseEntry().setQuantifiableComposition();
     }
 
-    public void setQuantity( int quantity ) {
-        getShowcaseEntry().getCraftable().setQuantity( quantity );
+    public void setQuantity(int quantity) {
+        getShowcaseEntry().getCraftable().setQuantity(quantity);
     }
 
     public void increaseQuantity() {
         getShowcaseEntry().getCraftable().increaseQuantity();
     }
 
-    public void increaseQuantity( int increment ) {
-        getShowcaseEntry().getCraftable().increaseQuantity( increment );
+    public void increaseQuantity(int increment) {
+        getShowcaseEntry().getCraftable().increaseQuantity(increment);
     }
 
     public void decreaseQuantity() {
         getShowcaseEntry().getCraftable().decreaseQuantity();
     }
 
-    public void decreaseQuantity( int decrement ) {
-        getShowcaseEntry().getCraftable().decreaseQuantity( decrement );
+    public void decreaseQuantity(int decrement) {
+        getShowcaseEntry().getCraftable().decreaseQuantity(decrement);
     }
 
-    private Cursor query( Context context, Uri uri ) {
-        return context.getContentResolver().query( uri, null, null, null, null );
+    private Cursor query(Context context, Uri uri) {
+        return context.getContentResolver().query(uri, null, null, null, null);
     }
 
-    private ShowcaseEntry QueryForDetails( Context context, long dlc_id, long _id, int quantity )
+    private ShowcaseEntry QueryForDetails(Context context, long dlc_id, long _id, int quantity)
             throws ExceptionUtil.CursorNullException, ExceptionUtil.CursorEmptyException {
-        Uri uri = DatabaseContract.EngramEntry.buildUriWithId( dlc_id, _id );
+        Uri uri = DatabaseContract.EngramEntry.buildUriWithId(dlc_id, _id);
 
-        try ( Cursor cursor = query( context, uri ) ) {
-            if ( cursor == null )
-                throw new ExceptionUtil.CursorNullException( uri );
+        try (Cursor cursor = query(context, uri)) {
+            if (cursor == null)
+                throw new ExceptionUtil.CursorNullException(uri);
 
-            if ( !cursor.moveToFirst() )
-                throw new ExceptionUtil.CursorEmptyException( uri );
+            if (!cursor.moveToFirst())
+                throw new ExceptionUtil.CursorEmptyException(uri);
 
-            String name = cursor.getString( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_NAME ) );
+            String name = cursor.getString(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_NAME));
 
-            String folder = cursor.getString( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_IMAGE_FOLDER ) );
-            String file = cursor.getString( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_IMAGE_FILE ) );
+            String folder = cursor.getString(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_IMAGE_FOLDER));
+            String file = cursor.getString(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_IMAGE_FILE));
 
-            String description = cursor.getString( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_DESCRIPTION ) );
-            int yield = cursor.getInt( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_YIELD ) );
-            long categoryId = cursor.getLong( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_CATEGORY_KEY ) );
-            int requiredLevel = cursor.getInt( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_LEVEL ) );
+            String description = cursor.getString(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_DESCRIPTION));
+            int yield = cursor.getInt(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_YIELD));
+            long categoryId = cursor.getLong(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_CATEGORY_KEY));
+            int requiredLevel = cursor.getInt(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_LEVEL));
 
-            List<Long> stationIds = new ArrayList<>( 0 );
-            stationIds.add( cursor.getLong( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_STATION_KEY ) ) );
-            while ( cursor.moveToNext() )
-                stationIds.add( cursor.getLong( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_STATION_KEY ) ) );
+            List<Long> stationIds = new ArrayList<>(0);
+            stationIds.add(cursor.getLong(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_STATION_KEY)));
+            while (cursor.moveToNext())
+                stationIds.add(cursor.getLong(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_STATION_KEY)));
 
             // Next, let's grab station details from ids
-            LongSparseArray<Station> stations = QueryForStations( context, dlc_id, stationIds );
+            LongSparseArray<Station> stations = QueryForStations(context, dlc_id, stationIds);
 
             // Next, let's grab matching Queue details, if there are any.
-            if ( quantity == 0 )
-                if ( CraftingQueue.getInstance().contains( _id ) )
-                    quantity = CraftingQueue.getInstance().getCraftable( _id ).getQuantity();
+            if (quantity == 0)
+                if (CraftingQueue.getInstance().contains(_id))
+                    quantity = CraftingQueue.getInstance().getCraftable(_id).getQuantity();
 
             // Finally, let's grab Composition details.
-            LongSparseArray<CompositeResource> composition = QueryForComposition( context, dlc_id, _id );
+            LongSparseArray<CompositeResource> composition = QueryForComposition(context, dlc_id, _id);
 
-            DisplayEngram craftable = new DisplayEngram( _id, name, folder, file, yield, categoryId, quantity );
+            DisplayEngram craftable = new DisplayEngram(_id, name, folder, file, yield, categoryId, quantity);
 
-            return new ShowcaseEntry( craftable, description, requiredLevel, dlc_id, composition, stations );
+            return new ShowcaseEntry(craftable, description, requiredLevel, dlc_id, composition, stations);
         }
     }
 
-    private LongSparseArray<Station> QueryForStations( Context context, long dlc_id, List<Long> ids )
+    private LongSparseArray<Station> QueryForStations(Context context, long dlc_id, List<Long> ids)
             throws ExceptionUtil.CursorNullException, ExceptionUtil.CursorEmptyException {
-        LongSparseArray<Station> stations = new LongSparseArray<>( 0 );
+        LongSparseArray<Station> stations = new LongSparseArray<>(0);
 
-        for ( long _id : ids ) {
-            Uri uri = DatabaseContract.StationEntry.buildUriWithId( dlc_id, _id );
+        for (long _id : ids) {
+            Uri uri = DatabaseContract.StationEntry.buildUriWithId(dlc_id, _id);
 
-            try ( Cursor cursor = query( context, uri ) ) {
-                if ( cursor == null )
-                    throw new ExceptionUtil.CursorNullException( uri );
+            try (Cursor cursor = query(context, uri)) {
+                if (cursor == null)
+                    throw new ExceptionUtil.CursorNullException(uri);
 
-                if ( !cursor.moveToFirst() )
-                    throw new ExceptionUtil.CursorEmptyException( uri );
+                if (!cursor.moveToFirst())
+                    throw new ExceptionUtil.CursorEmptyException(uri);
 
-                String name = cursor.getString( cursor.getColumnIndex( DatabaseContract.StationEntry.COLUMN_NAME ) );
-                String file = cursor.getString( cursor.getColumnIndex( DatabaseContract.StationEntry.COLUMN_IMAGE_FILE ) );
-                String folder = cursor.getString( cursor.getColumnIndex( DatabaseContract.StationEntry.COLUMN_IMAGE_FOLDER ) );
+                String name = cursor.getString(cursor.getColumnIndex(DatabaseContract.StationEntry.COLUMN_NAME));
+                String file = cursor.getString(cursor.getColumnIndex(DatabaseContract.StationEntry.COLUMN_IMAGE_FILE));
+                String folder = cursor.getString(cursor.getColumnIndex(DatabaseContract.StationEntry.COLUMN_IMAGE_FOLDER));
 
-                stations.put( _id, new Station( _id, name, folder, file ) );
+                stations.put(_id, new Station(_id, name, folder, file));
             }
         }
 
         return stations;
     }
 
-    private LongSparseArray<CompositeResource> QueryForComposition( Context context, long dlc_id, long engram_id )
+    private LongSparseArray<CompositeResource> QueryForComposition(Context context, long dlc_id, long engram_id)
             throws ExceptionUtil.CursorNullException, ExceptionUtil.CursorEmptyException {
-        Uri uri = DatabaseContract.CompositionEntry.buildUriWithEngramId( dlc_id, engram_id );
-        try ( Cursor cursor = context.getContentResolver().query( uri, null, null, null, null ) ) {
-            if ( cursor == null )
-                throw new ExceptionUtil.CursorNullException( uri );
+        Uri uri = DatabaseContract.CompositionEntry.buildUriWithEngramId(dlc_id, engram_id);
+        try (Cursor cursor = context.getContentResolver().query(uri, null, null, null, null)) {
+            if (cursor == null)
+                throw new ExceptionUtil.CursorNullException(uri);
 
-            LongSparseArray<CompositeResource> composition = new LongSparseArray<>( 0 );
-            while ( cursor.moveToNext() ) {
-                long _id = cursor.getLong( cursor.getColumnIndex( DatabaseContract.CompositionEntry.COLUMN_RESOURCE_KEY ) );
-                int quantity = cursor.getInt( cursor.getColumnIndex( DatabaseContract.CompositionEntry.COLUMN_QUANTITY ) );
+            LongSparseArray<CompositeResource> composition = new LongSparseArray<>(0);
+            while (cursor.moveToNext()) {
+                long _id = cursor.getLong(cursor.getColumnIndex(DatabaseContract.CompositionEntry.COLUMN_RESOURCE_KEY));
+                int quantity = cursor.getInt(cursor.getColumnIndex(DatabaseContract.CompositionEntry.COLUMN_QUANTITY));
 
-                composition.put( _id, new CompositeResource( QueryForResource( context, dlc_id, _id ), quantity ) );
+                composition.put(_id, new CompositeResource(QueryForResource(context, dlc_id, _id), quantity));
             }
 
             return composition;
         }
     }
 
-    private Resource QueryForResource( Context context, long dlc_id, long _id )
+    private Resource QueryForResource(Context context, long dlc_id, long _id)
             throws ExceptionUtil.CursorNullException, ExceptionUtil.CursorEmptyException {
-        Uri uri = DatabaseContract.ResourceEntry.buildUriWithId( dlc_id, _id );
+        Uri uri = DatabaseContract.ResourceEntry.buildUriWithId(dlc_id, _id);
 
-        try ( Cursor cursor = context.getContentResolver().query( uri, null, null, null, null ) ) {
-            if ( cursor == null )
-                throw new ExceptionUtil.CursorNullException( uri );
+        try (Cursor cursor = context.getContentResolver().query(uri, null, null, null, null)) {
+            if (cursor == null)
+                throw new ExceptionUtil.CursorNullException(uri);
 
-            if ( !cursor.moveToFirst() )
-                throw new ExceptionUtil.CursorEmptyException( uri );
+            if (!cursor.moveToFirst())
+                throw new ExceptionUtil.CursorEmptyException(uri);
 
-            String name = cursor.getString( cursor.getColumnIndex( DatabaseContract.ResourceEntry.COLUMN_NAME ) );
-            String folder = cursor.getString( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_IMAGE_FOLDER ) );
-            String file = cursor.getString( cursor.getColumnIndex( DatabaseContract.EngramEntry.COLUMN_IMAGE_FILE ) );
+            String name = cursor.getString(cursor.getColumnIndex(DatabaseContract.ResourceEntry.COLUMN_NAME));
+            String folder = cursor.getString(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_IMAGE_FOLDER));
+            String file = cursor.getString(cursor.getColumnIndex(DatabaseContract.EngramEntry.COLUMN_IMAGE_FILE));
 
-            return new Resource( _id, name, folder, file );
+            return new Resource(_id, name, folder, file);
         }
     }
 
-    private Category QueryForCategoryDetails( Context context, long dlc_id, long _id )
+    private Category QueryForCategoryDetails(Context context, long dlc_id, long _id)
             throws ExceptionUtil.CursorNullException, ExceptionUtil.CursorEmptyException {
-        Uri uri = DatabaseContract.CategoryEntry.buildUriWithId( dlc_id, _id );
+        Uri uri = DatabaseContract.CategoryEntry.buildUriWithId(dlc_id, _id);
 
-        try ( Cursor cursor = query( context, uri ) ) {
-            if ( cursor == null )
-                throw new ExceptionUtil.CursorNullException( uri );
+        try (Cursor cursor = query(context, uri)) {
+            if (cursor == null)
+                throw new ExceptionUtil.CursorNullException(uri);
 
-            if ( !cursor.moveToFirst() )
-                throw new ExceptionUtil.CursorEmptyException( uri );
+            if (!cursor.moveToFirst())
+                throw new ExceptionUtil.CursorEmptyException(uri);
 
-            String name = cursor.getString( cursor.getColumnIndex( DatabaseContract.CategoryEntry.COLUMN_NAME ) );
-            long parent_id = cursor.getLong( cursor.getColumnIndex( DatabaseContract.CategoryEntry.COLUMN_PARENT_KEY ) );
+            String name = cursor.getString(cursor.getColumnIndex(DatabaseContract.CategoryEntry.COLUMN_NAME));
+            long parent_id = cursor.getLong(cursor.getColumnIndex(DatabaseContract.CategoryEntry.COLUMN_PARENT_KEY));
 
-            return new Category( _id, name, parent_id );
+            return new Category(_id, name, parent_id);
         }
     }
 
-    private String QueryForDLCName( Context context, long dlc_id )
+    private String QueryForDLCName(Context context, long dlc_id)
             throws ExceptionUtil.CursorNullException, ExceptionUtil.CursorEmptyException {
-        Uri uri = DatabaseContract.buildUriWithId( DatabaseContract.DLCEntry.CONTENT_URI, dlc_id );
+        Uri uri = DatabaseContract.buildUriWithId(DatabaseContract.DLCEntry.CONTENT_URI, dlc_id);
 
-        try ( Cursor cursor = query( context, uri ) ) {
-            if ( cursor == null )
-                throw new ExceptionUtil.CursorNullException( uri );
+        try (Cursor cursor = query(context, uri)) {
+            if (cursor == null)
+                throw new ExceptionUtil.CursorNullException(uri);
 
-            if ( !cursor.moveToFirst() )
-                throw new ExceptionUtil.CursorEmptyException( uri );
+            if (!cursor.moveToFirst())
+                throw new ExceptionUtil.CursorEmptyException(uri);
 
-            return cursor.getString( cursor.getColumnIndex( DatabaseContract.DLCEntry.COLUMN_NAME ) );
+            return cursor.getString(cursor.getColumnIndex(DatabaseContract.DLCEntry.COLUMN_NAME));
         }
     }
 
@@ -329,37 +330,37 @@ public class Showcase {
         private LongSparseArray<QueueResource> mQuantifiableComposition;
         private LongSparseArray<Station> mStations;
 
-        public ShowcaseEntry( DisplayEngram craftable, String description, int requiredLevel, long dlc_id,
-                              LongSparseArray<CompositeResource> composition, LongSparseArray<Station> stations ) {
-            setCraftable( craftable );
+        ShowcaseEntry(DisplayEngram craftable, String description, int requiredLevel, long dlc_id,
+                      LongSparseArray<CompositeResource> composition, LongSparseArray<Station> stations) {
+            setCraftable(craftable);
 
-            setDescription( description );
-            setRequiredLevel( requiredLevel );
-            setDLCId( dlc_id );
+            setDescription(description);
+            setRequiredLevel(requiredLevel);
+            setDLCId(dlc_id);
 
-            setComposition( composition );
+            setComposition(composition);
             setQuantifiableComposition();
 
-            setStations( stations );
+            setStations(stations);
         }
 
-        void setCraftable( DisplayEngram craftable ) {
+        void setCraftable(DisplayEngram craftable) {
             this.mCraftable = craftable;
         }
 
-        void setDescription( String description ) {
+        void setDescription(String description) {
             this.mDescription = description;
         }
 
-        void setRequiredLevel( int requiredLevel ) {
+        void setRequiredLevel(int requiredLevel) {
             this.mRequiredLevel = requiredLevel;
         }
 
-        void setDLCId( long dlc_id ) {
+        void setDLCId(long dlc_id) {
             this.mDLCId = dlc_id;
         }
 
-        public void setComposition( LongSparseArray<CompositeResource> composition ) {
+        public void setComposition(LongSparseArray<CompositeResource> composition) {
             this.mComposition = composition;
         }
 
@@ -368,21 +369,21 @@ public class Showcase {
             int quantity = getCraftable().getQuantity();
 
             LongSparseArray<QueueResource> returnableComposition = new LongSparseArray<>();
-            for ( int i = 0; i < baseComposition.size(); i++ ) {
-                CompositeResource resource = baseComposition.valueAt( i );
+            for (int i = 0; i < baseComposition.size(); i++) {
+                CompositeResource resource = baseComposition.valueAt(i);
 
-                QueueResource queueResource = new QueueResource( resource, quantity );
-                returnableComposition.append( i, queueResource );
+                QueueResource queueResource = new QueueResource(resource, quantity);
+                returnableComposition.append(i, queueResource);
             }
 
-            mQuantifiableComposition = Util.sortResourcesByName( returnableComposition );
+            mQuantifiableComposition = Util.sortResourcesByName(returnableComposition);
         }
 
-        public void setStations( LongSparseArray<Station> stations ) {
+        public void setStations(LongSparseArray<Station> stations) {
             this.mStations = stations;
         }
 
-        public DisplayEngram getCraftable() {
+        DisplayEngram getCraftable() {
             return mCraftable;
         }
 
@@ -390,11 +391,11 @@ public class Showcase {
             return mDescription;
         }
 
-        public int getRequiredLevel() {
+        int getRequiredLevel() {
             return mRequiredLevel;
         }
 
-        public long getDLCId() {
+        long getDLCId() {
             return mDLCId;
         }
 
@@ -406,16 +407,16 @@ public class Showcase {
             return mQuantifiableComposition;
         }
 
-        public QueueResource getResource( int position ) throws Exception {
-            if ( Util.isValidPosition( position, getQuantifiableComposition().size() ) ) {
-                QueueResource resource = getQuantifiableComposition().valueAt( position );
+        QueueResource getResource(int position) throws Exception {
+            if (Util.isValidPosition(position, getQuantifiableComposition().size())) {
+                QueueResource resource = getQuantifiableComposition().valueAt(position);
 
-                if ( resource == null )
-                    throw new ExceptionUtil.ArrayElementNullException( position, getQuantifiableComposition().toString() );
+                if (resource == null)
+                    throw new ExceptionUtil.ArrayElementNullException(position, getQuantifiableComposition().toString());
 
                 return resource;
             } else {
-                throw new ExceptionUtil.PositionOutOfBoundsException( position, getQuantifiableComposition().size(), getQuantifiableComposition().toString() );
+                throw new ExceptionUtil.PositionOutOfBoundsException(position, getQuantifiableComposition().size(), getQuantifiableComposition().toString());
             }
         }
 
