@@ -1,4 +1,22 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.model;
+
+import androidx.annotation.NonNull;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,53 +32,53 @@ public abstract class SortableMap {
     private int mSize;
 
     public SortableMap() {
-        mKeys = new ArrayList<>( 0 );
-        mValues = new ArrayList<>( 0 );
+        mKeys = new ArrayList<>(0);
+        mValues = new ArrayList<>(0);
         mSize = 0;
     }
 
-    public Object get( long key ) {
-        return get( key, null );
+    public Object get(long key) {
+        return get(key, null);
     }
 
-    public Object get( long key, Object valueIfNotFound ) {
-        if ( contains( key ) ) {
-            int position = mKeys.indexOf( key );
-            return valueAt( position );
+    public Object get(long key, Object valueIfNotFound) {
+        if (contains(key)) {
+            int position = mKeys.indexOf(key);
+            return valueAt(position);
         } else {
             return valueIfNotFound;
         }
     }
 
-    public boolean contains( long key ) {
-        return mKeys.contains( key );
+    public boolean contains(long key) {
+        return mKeys.contains(key);
     }
 
-    public int indexOfKey( long key ) {
-        return mKeys.indexOf( key );
+    public int indexOfKey(long key) {
+        return mKeys.indexOf(key);
     }
 
-    public long keyAt( int position ) {
-        return mKeys.get( position );
+    private long keyAt(int position) {
+        return mKeys.get(position);
     }
 
-    public Object valueAt( int position ) {
+    public Object valueAt(int position) {
         try {
-            return mValues.get( position );
-        } catch ( IndexOutOfBoundsException e ) {
-            ExceptionObserver.getInstance().notifyFatalExceptionCaught( TAG, new
-                    ExceptionUtil.PositionOutOfBoundsException( position, mValues.size(), mValues.toString() ) );
+            return mValues.get(position);
+        } catch (IndexOutOfBoundsException e) {
+            ExceptionObserver.getInstance().notifyFatalExceptionCaught(TAG, new
+                    ExceptionUtil.PositionOutOfBoundsException(position, mValues.size(), mValues.toString()));
             return null;
         }
     }
 
-    public void setValueAt( int position, Object value ) {
-        set( position, keyAt( position ), value );
+    public void setValueAt(int position, Object value) {
+        set(position, keyAt(position), value);
     }
 
-    public void removeAt( int position ) {
-        mKeys.remove( position );
-        mValues.remove( position );
+    void removeAt(int position) {
+        mKeys.remove(position);
+        mValues.remove(position);
         mSize--;
     }
 
@@ -80,28 +98,29 @@ public abstract class SortableMap {
      * @param key   long value key taken from Object's _id (ROWID)
      * @param value value that will be added to lists
      */
-    public void add( long key, Object value ) {
-        if ( !contains( key ) ) {
-            mKeys.add( key );
-            mValues.add( value );
+    public void add(long key, Object value) {
+        if (!contains(key)) {
+            mKeys.add(key);
+            mValues.add(value);
             mSize++;
         }
     }
 
     /**
      * Public method that, no matter if key exists, will add a new object to the lists.
+     *
      * @param key   long value key taken from Object's _id (ROWID)
      * @param value value that will be added to lists
      */
-    public void addNew( long key, Object value ) {
-        mKeys.add( key );
-        mValues.add( value );
+    public void addNew(long key, Object value) {
+        mKeys.add(key);
+        mValues.add(value);
         mSize++;
     }
 
-    public void addAll( SortableMap map ) {
-        mKeys.addAll( map.keySet() );
-        mValues.addAll( map.valueSet() );
+    public void addAll(SortableMap map) {
+        mKeys.addAll(map.keySet());
+        mValues.addAll(map.valueSet());
         mSize += map.size();
     }
 
@@ -111,13 +130,13 @@ public abstract class SortableMap {
      * @param key   long value key taken from Object's _id (ROWID)
      * @param value value that will be added to lists
      */
-    public void put( long key, Object value ) {
-        if ( contains( key ) ) {
-            int position = indexOfKey( key );
+    public void put(long key, Object value) {
+        if (contains(key)) {
+            int position = indexOfKey(key);
 
-            set( position, key, value );
+            set(position, key, value);
         } else {
-            add( key, value );
+            add(key, value);
         }
     }
 
@@ -129,9 +148,9 @@ public abstract class SortableMap {
         return mValues;
     }
 
-    private void set( int position, long key, Object value ) {
-        mKeys.set( position, key );
-        mValues.set( position, value );
+    private void set(int position, long key, Object value) {
+        mKeys.set(position, key);
+        mValues.set(position, value);
     }
 
     /**
@@ -140,25 +159,26 @@ public abstract class SortableMap {
      * @param position position of Object in map
      * @return Comparable object, be it String or whatever, used to help sort().
      */
-    public abstract Comparable getComparable( int position );
+    public abstract Comparable getComparable(int position);
 
-    private void swap( int a, int b ) {
-        long tempKey = keyAt( a );
-        Object tempValue = valueAt( a );
-        set( a, keyAt( b ), valueAt( b ) );
-        set( b, tempKey, tempValue );
+    private void swap(int a, int b) {
+        long tempKey = keyAt(a);
+        Object tempValue = valueAt(a);
+        set(a, keyAt(b), valueAt(b));
+        set(b, tempKey, tempValue);
     }
 
     public void sort() {
 //        Log.d( TAG, "sort: before: " + toString() );
         boolean swapped = true;
-        while ( swapped ) {
+        while (swapped) {
             swapped = false;
-            for ( int i = 0; i < size() - 1; i++ ) {
-                Comparable first = getComparable( i );
-                Comparable second = getComparable( i + 1 );
-                if ( first.compareTo( second ) > 0 ) {
-                    swap( i, i + 1 );
+            for (int i = 0; i < size() - 1; i++) {
+                Comparable first = getComparable(i);
+                Comparable second = getComparable(i + 1);
+                //noinspection unchecked
+                if (first.compareTo(second) > 0) {
+                    swap(i, i + 1);
                     swapped = true;
                 }
             }
@@ -166,19 +186,20 @@ public abstract class SortableMap {
 //        Log.d( TAG, "sort: after: " + toString() );
     }
 
+    @NonNull
     @Override
     public String toString() {
         StringBuilder builder = new StringBuilder();
 
-        builder.append( '{' ).append( ' ' );
-        for ( int i = 0; i < size(); i++ ) {
-            builder.append( valueAt( i ).toString() );
-            if ( i < size() )
-                builder.append( ',' );
+        builder.append('{').append(' ');
+        for (int i = 0; i < size(); i++) {
+            builder.append(valueAt(i).toString());
+            if (i < size())
+                builder.append(',');
 
-            builder.append( ' ' );
+            builder.append(' ');
         }
-        builder.append( '}' );
+        builder.append('}');
 
         return builder.toString();
     }
