@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.util;
 
 import android.content.Context;
@@ -11,10 +27,10 @@ import java.util.Arrays;
 
 public class ExceptionUtil {
 
-    public static void SendErrorReportWithAlertDialog( final Context context, String tag, final Exception e ) {
-        SendErrorReport( tag, e );
+    public static void SendErrorReportWithAlertDialog(final Context context, String tag, final Exception e) {
+        SendErrorReport(tag, e);
 
-        DialogUtil.Error( context, new DialogUtil.Callback() {
+        DialogUtil.Error(context, new DialogUtil.Callback() {
             @Override
             public void onOk() {
                 // call reset to defaults!
@@ -23,97 +39,97 @@ public class ExceptionUtil {
 
             @Override
             public void onCancel() {
-                // forcily close app
-                throw new RuntimeException( "Force closing." );
+                // forcibly close app
+                throw new RuntimeException("Force closing.");
             }
-        } ).show();
+        }).show();
     }
 
-    public static void SendErrorReport( String tag, Exception e ) {
-        if ( e.getCause() != null )
-            FirebaseCrash.logcat( Log.ERROR, tag, e.getCause().toString() );
+    public static void SendErrorReport(String tag, Exception e) {
+        if (e.getCause() != null)
+            FirebaseCrash.logcat(Log.ERROR, tag, e.getCause().toString());
 
-        FirebaseCrash.logcat( Log.ERROR, tag, e.toString() );
-        FirebaseCrash.logcat( Log.ERROR, tag, e.getMessage() );
-        FirebaseCrash.logcat( Log.ERROR, tag, Arrays.toString( e.getStackTrace() ) );
-        FirebaseCrash.logcat( Log.ERROR, tag, BuildErrorReportPreferencesBundle().toString() );
+        FirebaseCrash.logcat(Log.ERROR, tag, e.toString());
+        FirebaseCrash.logcat(Log.ERROR, tag, e.getMessage());
+        FirebaseCrash.logcat(Log.ERROR, tag, Arrays.toString(e.getStackTrace()));
+//        FirebaseCrash.logcat(Log.ERROR, tag, BuildErrorReportPreferencesBundle().toString());
 
-        FirebaseCrash.report( e );
+        FirebaseCrash.report(e);
     }
 
     // if position of requested list is out of bounds, then throw this exception
     public static class PositionOutOfBoundsException extends Exception {
-        public PositionOutOfBoundsException( int index, int size, String contents ) {
-            super( BuildExceptionMessageBundle( index, size, contents ).toString() );
+        public PositionOutOfBoundsException(int index, int size, String contents) {
+            super(BuildExceptionMessageBundle(index, size, contents).toString());
         }
     }
 
     // if list<object> element = null, then throw this exception
     public static class ArrayElementNullException extends Exception {
-        public ArrayElementNullException( int index, String contents ) {
-            super( BuildExceptionMessageBundle( index, contents ).toString() );
+        public ArrayElementNullException(int index, String contents) {
+            super(BuildExceptionMessageBundle(index, contents).toString());
         }
     }
 
     // Uri given to ContentProvider does not match uri list
     public static class URIUnknownException extends Exception {
-        public URIUnknownException( Uri uri ) {
-            super( uri.toString() );
+        public URIUnknownException(Uri uri) {
+            super(uri.toString());
         }
     }
 
     // Cursor returned null, onResume requested _id
     public static class CursorNullException extends Exception {
 
-        public CursorNullException( Uri uri ) {
-            super( uri.toString() );
+        public CursorNullException(Uri uri) {
+            super(uri.toString());
         }
 
-        public CursorNullException( Uri uri, String contents ) {
-            super( uri.toString() + " array:" + contents );
+        public CursorNullException(Uri uri, String contents) {
+            super(uri.toString() + " array:" + contents);
         }
     }
 
     // Cursor returned null, onResume requested _id
     public static class CursorEmptyException extends Exception {
-        public CursorEmptyException( Uri uri ) {
-            super( uri.toString() );
+        public CursorEmptyException(Uri uri) {
+            super(uri.toString());
         }
     }
 
     // Bundle used to express an array's contents and the requested index
-    private static Bundle BuildExceptionMessageBundle( int index, String contents ) {
+    private static Bundle BuildExceptionMessageBundle(int index, String contents) {
         Bundle bundle = new Bundle();
 
-        bundle.putInt( "array_index", index );
-        bundle.putString( "array_contents", contents );
+        bundle.putInt("array_index", index);
+        bundle.putString("array_contents", contents);
 
         return bundle;
     }
 
-    private static Bundle BuildExceptionMessageBundle( int index, int size, String contents ) {
+    private static Bundle BuildExceptionMessageBundle(int index, int size, String contents) {
         Bundle bundle = new Bundle();
 
-        bundle.putInt( "array_index", index );
-        bundle.putInt( "array_size", size );
-        bundle.putString( "array_contents", contents );
+        bundle.putInt("array_index", index);
+        bundle.putInt("array_size", size);
+        bundle.putString("array_contents", contents);
 
         return bundle;
     }
 
-    private static Bundle BuildErrorReportPreferencesBundle() {
+    private static Bundle BuildErrorReportPreferencesBundle(Context context) {
         try {
-//            PrefsUtil prefs = PrefsUtil.getInstance();
+            PrefsUtil prefs = PrefsUtil.getInstance(context);
 
             Bundle bundle = new Bundle();
-//            bundle.putBoolean( "filter_by_category", prefs.getCategoryFilterPreference() );
-//            bundle.putBoolean( "filter_by_station", prefs.getStationFilterPreference() );
-//            bundle.putBoolean( "filter_by_level", prefs.getLevelFilterPreference() );
-//            bundle.putInt( "current_level", prefs.getRequiredLevel() );
-//            bundle.putBoolean( "breakdown_resources", prefs.getRefinedFilterPreference() );
+            bundle.putBoolean("filter_by_category", prefs.getCategoryFilterPreference());
+            bundle.putBoolean("filter_by_station", prefs.getStationFilterPreference());
+            bundle.putBoolean("filter_by_level", prefs.getLevelFilterPreference());
+            bundle.putInt("current_level", prefs.getRequiredLevel());
+            bundle.putBoolean("breakdown_resources", prefs.getRefinedFilterPreference());
 
             return bundle;
-        } catch ( Exception e ) {
+        } catch (Exception e) {
             return new Bundle();
         }
     }
