@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.util;
 
 import android.os.AsyncTask;
@@ -11,7 +27,7 @@ class IssueUtil {
 
     private static class IssueTask extends AsyncTask<Void, Void, Void> {
         @Override
-        protected Void doInBackground( Void... params ) {
+        protected Void doInBackground(Void... params) {
 //            try {
 //                Log.d( TAG, "issueTask" );
 //
