@@ -1,11 +1,28 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.views;
 
 import android.content.Context;
 import android.os.Build;
-import android.support.annotation.Nullable;
-import android.support.annotation.RequiresApi;
 import android.util.AttributeSet;
 import android.widget.LinearLayout;
+
+import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 
 import arc.resource.calculator.R;
 
@@ -13,28 +30,28 @@ public class CraftableLayout extends LinearLayout {
     CraftableSwitcher mSwitcher;
     NavigationTextView mTextView;
 
-    public CraftableLayout( Context context ) {
-        super( context );
+    public CraftableLayout(Context context) {
+        super(context);
     }
 
-    public CraftableLayout( Context context, @Nullable AttributeSet attrs ) {
-        super( context, attrs );
+    public CraftableLayout(Context context, @Nullable AttributeSet attrs) {
+        super(context, attrs);
     }
 
-    public CraftableLayout( Context context, @Nullable AttributeSet attrs, int defStyleAttr ) {
-        super( context, attrs, defStyleAttr );
+    public CraftableLayout(Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
     }
 
-    @RequiresApi( api = Build.VERSION_CODES.LOLLIPOP )
-    public CraftableLayout( Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes ) {
-        super( context, attrs, defStyleAttr, defStyleRes );
+    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    public CraftableLayout(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
     }
 
     public void onCreate() {
-        mSwitcher = ( CraftableSwitcher ) findViewById( R.id.switcher_craftables );
+        mSwitcher = findViewById(R.id.switcher_craftables);
         mSwitcher.onCreate();
 
-        mTextView = ( NavigationTextView ) findViewById( R.id.textview_navigation_hierarchy );
+        mTextView = findViewById(R.id.textview_navigation_hierarchy);
         mTextView.create();
     }
 
@@ -53,7 +70,7 @@ public class CraftableLayout extends LinearLayout {
         mSwitcher.onDestroy();
     }
 
-    public void onSearch( String query ) {
-        mSwitcher.onSearch( query );
+    public void onSearch(String query) {
+        mSwitcher.onSearch(query);
     }
 }
