@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.views;
 
 import android.content.Context;
@@ -15,54 +31,54 @@ public class CraftableSwitcher extends ViewSwitcher implements CraftableRecycler
     private CraftableRecyclerView mRecyclerView;
 
     @Override
-    public void onError( Exception e ) {
-        if ( !isTextViewShown() )
+    public void onError(Exception e) {
+        if (!isTextViewShown())
             showNext();
 
-        onStatusUpdate( "An error occurred while fetching Engram data." );
+        onStatusUpdate("An error occurred while fetching Engram data.");
 
-        ExceptionObserver.getInstance().notifyExceptionCaught( TAG, e );
+        ExceptionObserver.getInstance().notifyExceptionCaught(TAG, e);
     }
 
     @Override
     public void onInit() {
-        if ( !isTextViewShown() )
+        if (!isTextViewShown())
             showNext();
 
-        onStatusUpdate( "Fetching Engram data.." );
+        onStatusUpdate("Fetching Engram data..");
     }
 
     @Override
     public void onPopulated() {
         // switch view to recycler
-        if ( !isRecyclerViewShown() )
+        if (!isRecyclerViewShown())
             showNext();
     }
 
     @Override
     public void onEmpty() {
         // switch view to textview, display empty message
-        if ( !isTextViewShown() )
+        if (!isTextViewShown())
             showNext();
 
-        onStatusUpdate( "Engram data is empty." );
+        onStatusUpdate("Engram data is empty.");
     }
 
-    public CraftableSwitcher( Context context ) {
-        super( context );
+    public CraftableSwitcher(Context context) {
+        super(context);
     }
 
-    public CraftableSwitcher( Context context, AttributeSet attrs ) {
-        super( context, attrs );
+    public CraftableSwitcher(Context context, AttributeSet attrs) {
+        super(context, attrs);
     }
 
     public void onCreate() {
         // instantiate our textview for status updates
-        mTextView = ( TextView ) findViewById( R.id.textview_craftables );
+        mTextView = findViewById(R.id.textview_craftables);
 
         // instantiate recyclerView
-        mRecyclerView = ( CraftableRecyclerView ) findViewById( R.id.gridview_craftables );
-        mRecyclerView.create( this );
+        mRecyclerView = findViewById(R.id.gridview_craftables);
+        mRecyclerView.create(this);
     }
 
     public void onResume() {
@@ -77,12 +93,12 @@ public class CraftableSwitcher extends ViewSwitcher implements CraftableRecycler
         mRecyclerView.destroy();
     }
 
-    public void onSearch( String query ) {
+    public void onSearch(String query) {
         mRecyclerView.search(query);
     }
 
-    private void onStatusUpdate( String text ) {
-        mTextView.setText( text );
+    private void onStatusUpdate(String text) {
+        mTextView.setText(text);
     }
 
     private boolean isTextViewShown() {
