@@ -1,8 +1,25 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.views;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatTextView;
 import android.util.AttributeSet;
+
+import androidx.appcompat.widget.AppCompatTextView;
 
 import arc.resource.calculator.listeners.NavigationObserver;
 
@@ -11,25 +28,25 @@ public class NavigationTextView extends AppCompatTextView {
 
     private NavigationObserver.Listener mListener = new NavigationObserver.Listener() {
         @Override
-        public void onUpdate( String text ) {
-            setText( text );
+        public void onUpdate(String text) {
+            setText(text);
         }
     };
 
-    public NavigationTextView( Context context ) {
-        super( context );
+    public NavigationTextView(Context context) {
+        super(context);
     }
 
-    public NavigationTextView( Context context, AttributeSet attrs ) {
-        super( context, attrs );
+    public NavigationTextView(Context context, AttributeSet attrs) {
+        super(context, attrs);
     }
 
-    public NavigationTextView( Context context, AttributeSet attrs, int defStyleAttr ) {
-        super( context, attrs, defStyleAttr );
+    public NavigationTextView(Context context, AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
     }
 
     public void create() {
-        NavigationObserver.getInstance().registerListener( mListener );
+        NavigationObserver.getInstance().registerListener(mListener);
     }
 
     public void resume() {
@@ -39,6 +56,6 @@ public class NavigationTextView extends AppCompatTextView {
     }
 
     public void destroy() {
-        NavigationObserver.getInstance().unregisterListener( mListener );
+        NavigationObserver.getInstance().unregisterListener(mListener);
     }
 }
