@@ -1,41 +1,53 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.views;
 
 import android.content.Context;
-import android.support.v7.widget.AppCompatCheckBox;
-import android.support.v7.widget.AppCompatTextView;
 import android.util.AttributeSet;
 import android.view.View;
 
-import arc.resource.calculator.listeners.NavigationObserver;
-import arc.resource.calculator.listeners.QueueObserver;
-import arc.resource.calculator.model.CraftingQueue;
-import arc.resource.calculator.util.ListenerUtil;
+import androidx.appcompat.widget.AppCompatCheckBox;
+
 import arc.resource.calculator.util.PrefsUtil;
 
 public class RawMaterialCheckBox extends AppCompatCheckBox {
     private static final String TAG = RawMaterialCheckBox.class.getSimpleName();
 
-    public RawMaterialCheckBox( Context context ) {
-        super( context );
+    public RawMaterialCheckBox(Context context) {
+        super(context);
     }
 
-    public RawMaterialCheckBox( Context context, AttributeSet attrs ) {
-        super( context, attrs );
+    public RawMaterialCheckBox(Context context, AttributeSet attrs) {
+        super(context, attrs);
     }
 
-    public RawMaterialCheckBox( Context context, AttributeSet attrs, int defStyleAttr ) {
-        super( context, attrs, defStyleAttr );
+    public RawMaterialCheckBox(Context context, AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
     }
 
     public void create() {
-        setOnClickListener( new OnClickListener() {
+        setOnClickListener(new OnClickListener() {
             @Override
-            public void onClick( View v ) {
-                PrefsUtil.getInstance( getContext() ).saveRefinedFilterPreference( isChecked() );
+            public void onClick(View v) {
+                PrefsUtil.getInstance(getContext()).saveRefinedFilterPreference(isChecked());
             }
-        } );
+        });
 
-        setChecked( PrefsUtil.getInstance( getContext() ).getRefinedFilterPreference() );
+        setChecked(PrefsUtil.getInstance(getContext()).getRefinedFilterPreference());
     }
 
     public void resume() {
