@@ -1,35 +1,53 @@
+/*
+ * Copyright (c) 2019 Jared Stone
+ *
+ * This work is licensed under the Creative Commons
+ * Attribution-NonCommercial-NoDerivatives 4.0 International
+ * License. To view a copy of this license, visit
+ *
+ * http://creativecommons.org/licenses/by-nc-nd/4.0/
+ *
+ * or send a letter to
+ *
+ *  Creative Commons,
+ *  PO Box 1866,
+ *  Mountain View, CA 94042, USA.
+ */
+
 package arc.resource.calculator.views;
 
 import android.content.Context;
 import android.os.Build;
-import android.support.annotation.RequiresApi;
 import android.util.AttributeSet;
 import android.widget.RelativeLayout;
+
+import androidx.annotation.RequiresApi;
 
 /**
  * Created by jared on 4/27/2017.
  */
 
 public class SquareRelativeLayout extends RelativeLayout {
-    public SquareRelativeLayout( Context context ) {
-        super( context );
+    public SquareRelativeLayout(Context context) {
+        super(context);
     }
 
-    public SquareRelativeLayout( Context context, AttributeSet attrs ) {
-        super( context, attrs );
+    public SquareRelativeLayout(Context context, AttributeSet attrs) {
+        super(context, attrs);
     }
 
-    public SquareRelativeLayout( Context context, AttributeSet attrs, int defStyleAttr ) {
-        super( context, attrs, defStyleAttr );
+    public SquareRelativeLayout(Context context, AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
     }
 
-    @RequiresApi( api = Build.VERSION_CODES.LOLLIPOP )
-    public SquareRelativeLayout( Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes ) {
-        super( context, attrs, defStyleAttr, defStyleRes );
+    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    public SquareRelativeLayout(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
     }
 
     @Override
-    protected void onMeasure( int widthMeasureSpec, int heightMeasureSpec ) {
-        super.onMeasure( heightMeasureSpec, heightMeasureSpec );
+    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
+        //noinspection SuspiciousNameCombination
+        super.onMeasure(heightMeasureSpec, heightMeasureSpec);
     }
 }
